# [libbeat] Go Benchmarks comparing compress/gzip with klauspost/compress