--- a/libbeat/esleg/eslegclient/enc.go
+++ b/libbeat/esleg/eslegclient/enc.go
@@ -19,10 +19,11 @@ package eslegclient
 
 import (
 	"bytes"
-	"compress/gzip"
 	"io"
 	"net/http"
 	"time"
+
+	"github.com/klauspost/compress/gzip"
 
 	"github.com/elastic/beats/v7/libbeat/beat"
 	"github.com/elastic/beats/v7/libbeat/outputs/codec"