<!-- Type of change
Please label this PR with one of the following labels, depending on the scope of your change:
- Bug
- Enhancement
- Breaking change
- Deprecation
- Cleanup
- Docs
-->

## Proposed commit message

<!-- Mandatory
Explain here the changes you made on the PR.

Please explain:

- WHAT: patterns used, algorithms implemented, design architecture, message processing, etc.
- WHY:  the rationale/motivation for the changes

This text will be pasted into the squash dialog when the change is committed and will be
a long term historical record of the change to help future contributors understand the
change, please help them by making it clear and comprehensive, they may be you.

If the commit title is adequate to describe both of these things, The text here may be omitted
or replaced with "See title". The title of the PR will be used as the commit message title when
the merge is made and the "See title" marker will be removed if present.

The text here and the PR title will be subject to the PR review process.
-->
We send data to the `_bulk` endpoint and benchmark the results with and without the https://github.com/klauspost/compress library.

Results

```
goos: darwin
goarch: arm64
pkg: github.com/elastic/beats/v7/libbeat/outputs/elasticsearch
cpu: Apple M3 Pro
               │ without_klauspost_compress.txt │         with_klauspost.txt         │
               │             sec/op             │   sec/op     vs base               │
Compression-12                      50.41µ ± 7%   48.61µ ± 3%  -3.57% (p=0.002 n=10)

               │ without_klauspost_compress.txt │         with_klauspost.txt          │
               │              B/op              │     B/op      vs base               │
Compression-12                     10.70Ki ± 0%   10.76Ki ± 0%  +0.63% (p=0.001 n=10)

               │ without_klauspost_compress.txt │       with_klauspost.txt       │
               │           allocs/op            │ allocs/op   vs base            │
Compression-12                       130.0 ± 0%   130.0 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to have the PR approved

List here all the items you have verified BEFORE sending this PR. Please DO NOT remove any item, striking through those that do not apply. (Just in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have made corresponding change to the default configuration files
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.



## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
- Relates [#38599](https://github.com/elastic/beats/issues/38599)