@rdner 2024-11-13
@khushijain21 from the table looks like the compression level 4 always wins. I wonder what's the payload size difference between `compress/gzip` and `klauspost/compress` in this case?

I think it's important to compare encoded sizes, otherwise if `klauspost/compress` ~35% faster but the compressed payload is 50% larger we would slow down on networking more than we win on the CPU.